fix(app-control): read the AM refusal's demod rate only on radio builds
The width refusal compared a rate that radio-off builds always left at
0, so the check was dead code there. Keep the rate lookup and the check
under USE_RADIO. In the tests, include runtime/config.h for the mode
constants the CLI parse cases use, and index the AM impulse block with
size_t.

+++ tests/dsp/test_dsp_am_demod.cpp
@@ -16,6 +16,7 @@
  */
 
 #include <cmath>
+#include <cstddef>
 #include <cstdint>
 #include <cstdio>
 #include <dsd-neo/dsp/demod_pipeline.h>
@@ -339,7 +340,7 @@ test_floor_and_clamp(void) {
     rc |= expect("no carrier is not silence", silent);
 
     s->am_carrier = 0.5f;
-    for (int k = 0; k < 8; k++) {
+    for (size_t k = 0; k < 8; k++) {
         s->input_cb_buf[2 * k] = (k == 3) ? 20.0f : 0.5f;
         s->input_cb_buf[(2 * k) + 1] = 0.0f;
     }
